+++ npm-packages/private-demos/clerk-initial-auth/src/Badge.tsx
@@ -1,0 +1,11 @@
+import { useUser } from "@clerk/clerk-react";
+
+export default function Badge() {
+  const { user } = useUser();
+
+  return (
+    <p className="badge">
+      <span>Logged in{user!.fullName ? ` as ${user!.fullName}` : ""}</span>
+    </p>
+  );
+}