+++ npm-packages/convex/CHANGELOG.md
@@ -13,6 +13,14 @@
 - Fix a bug where an auth token passed initially to the ConvexHttpClient was
   ignored.
 
+- Experimental `expectAuth` option for Convex clients for indicating that
+  setAuth() will be called soon, so to wait for that token. Once setAuth() has
+  been called the existing token-waiting behavior takes over.
+
+  This is useful for applications that create a client and run a mutation,
+  query, or action _before_ setAuth() has been called, e.g. via a provider in
+  React.
+
 ## 1.25.4
 
 - Experimental `convex dev --once --debug-node-apis` debug flag for tracing
@@ -277,7 +285,6 @@ To upgrade to this release you'll need to upgrade any Convex components you use.
 
   because this pattern causes problems and there are straightforward
   workarounds. The problems here:
-
   1. Arguments and return values aren't validated despite the presence of
      validators at the function definition site.
   2. Functions called this way unexpectedly lack isolation and atomicity. Convex
@@ -287,7 +294,6 @@ To upgrade to this release you'll need to upgrade any Convex components you use.
      deadlocks and other bad behavior.
 
   There are two options for how to modify your code to address the warning:
-
   1. Refactor it out as a helper function, then call that helper function
      directly.
   2. Use `ctx.runMutation`, `ctx.runQuery`, or `ctx.runAction()` instead of
@@ -312,7 +318,6 @@ To upgrade to this release you'll need to upgrade any Convex components you use.
   there are many breaking changes.
 
 - Improvements to `npx convex run`:
-
   - Better argument parsing with JSON5 so `{ name: "sshader" }` parses
   - support for `--identity` similar to dashboard "acting as user" feature, like
     `npx convex run --identity '{ name: "sshader" }'`